+++ tools/integration_tests/read_large_files/seq_read_large_file_test.go
@@ -16,7 +16,6 @@ package read_large_files
 
 import (
 	"bytes"
-	"math/rand"
 	"os"
 	"path"
 	"testing"
@@ -25,33 +24,30 @@ import (
 	"github.com/googlecloudplatform/gcsfuse/tools/integration_tests/util/setup"
 )
 
-func TestReadLargeFileRandomly(t *testing.T) {
+func TestReadLargeFileSequentially(t *testing.T) {
 	// Clean the mountedDirectory before running test.
 	setup.CleanMntDir()
 
 	fileInLocalDisk := path.Join(os.Getenv("HOME"), FiveHundredMBFile)
 	file := path.Join(setup.MntDir(), FiveHundredMBFile)
 	// Create and copy the local file in mountedDirectory.
-	CreateFileInLocalDiskAndCopyFileFromLocalDiskToMntDir(fileInLocalDisk, file, FiveHundredMB, t)
+	CreateFileOnDiskAndCopyToMntDir(fileInLocalDisk, file, FiveHundredMB, t)
 
-	for i := 0; i < NumberOfRandomReadCalls; i++ {
-		offset := rand.Int63n(MaxReadableByteFromFile - MinReadableByteFromFile)
-		// Randomly read the data from file in mountedDirectory.
-		content, err := operations.ReadChunkFromFile(file, ChunkSize, offset)
-		if err != nil {
-			t.Errorf("Error in reading file: %v", err)
-		}
+	// Sequentially read the data from file.
+	content, err := operations.ReadFileSequentially(file, ChunkSize)
+	if err != nil {
+		t.Errorf("Error in reading file: %v", err)
+	}
 
-		// Read actual content from file located in local disk.
-		actualContent, err := operations.ReadChunkFromFile(fileInLocalDisk, ChunkSize, offset)
-		if err != nil {
-			t.Errorf("Error in reading file: %v", err)
-		}
+	// Read actual content from file located in local disk.
+	actualContent, err := operations.ReadFile(fileInLocalDisk)
+	if err != nil {
+		t.Errorf("Error in reading file: %v", err)
+	}
 
-		// Compare actual content and expect content.
-		if bytes.Equal(actualContent, content) == false {
-			t.Errorf("Error in reading file randomly.")
-		}
+	// Compare actual content and expect content.
+	if bytes.Equal(actualContent, content) == false {
+		t.Errorf("Error in reading file sequentially.")
 	}
 
 	// Removing file after testing.
